requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
jinja2==3.0.1
yt-dlp==2023.3.4
python-dotenv==0.19.0
websockets==10.0 